software/backend/requirements.txt
fastapi
uvicorn
paho-mqtt
pydantic
python-dotenv
